requirements.txt
fastapi==0.115.5
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.32.0
